requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
Jinja2>=3.1.4
